Create README.md
For college work I developed a food scale that sent food weight data via arduino and bluetooth to an android app and this data was stored in google firebase. Thus, the logged in user could save the weight of the food and also select foods already saved in firebase with their respective kcal, proteins, carbohydrates and fats per grams and set up their diet.